+++ spinta/components.py
@@ -38,6 +38,7 @@ if TYPE_CHECKING:
     from spinta.accesslog import AccessLog
     from spinta.formats.components import Format
     from spinta.dimensions.comments.components import Comment
+    from spinta.dimensions.scope.components import Scope
 
 
 class Context:
@@ -620,6 +621,14 @@ def page_in_data(data: dict) -> bool:
     return "_page" in data
 
 
+def revision_in_data(data: dict) -> bool:
+    return "_revision" in data
+
+
+def check_if_revision_explicit(prop: Property) -> bool:
+    return prop.name == "_revision" and prop.explicitly_given
+
+
 class ParamsPage:
     values: List[Any]
     size: int
@@ -642,6 +651,7 @@ class Model(MetaData):
     description: str
     ns: Namespace
     external: Entity = None
+    scopes: Dict[str, Scope]
     properties: Dict[str, Property]
     mode: Mode = None
     given: ModelGiven
@@ -668,6 +678,7 @@ class Model(MetaData):
         "unique": {"default": []},
         "base": {},
         "link": {},
+        "scopes": {"default": {}},
         "properties": {"default": {}},
         "external": {},
         "level": {
@@ -886,6 +897,7 @@ class Attachment:
 class UrlParseNode(TypedDict):
     name: str
     args: List[Any]
+    id_prop: Property
 
 
 class UrlParams: