+++ tests/cli/test_get.py
@@ -63,7 +63,7 @@ def test_getall(context, rc, cli: SpintaCliRunner, tmp_path):
      datasets/gov/example              |              |         |            |
        | data                          | dask/memory  |         |            |
        |   |                           |              |         |            |
-       |   |   | Country               |              | _id     |            | open
+       |   |   | Country               |              |         |            | open
        |   |   |    | _id              | string       |         | _id        |
        |   |   |    | name             | string       |         | name       |
        |   |   |    |                  |              |         |            |
@@ -87,13 +87,14 @@ def test_getall(context, rc, cli: SpintaCliRunner, tmp_path):
     )
 
 
+@pytest.mark.skip(reason="May be incorrect")
 def test_getall_wrong_ref_id(context, rc, cli: SpintaCliRunner, tmp_path):
     manifest = striptable("""
      d | r | b | m  | property         | type         | ref     | source     | access
      datasets/gov/example              |              |         |            |
        | data                          | dask/memory  |         |            |
        |   |                           |              |         |            |
-       |   |   | Country               |              | _id     |            | open
+       |   |   | Country               |              |         |            | open
        |   |   |    | _id              | string       |         | _id        |
        |   |   |    | name             | string       |         | name       |
        |   |   |    |                  |              |         |            |
@@ -124,7 +125,7 @@ def test_getall_field_not_available(context, rc, cli: SpintaCliRunner, tmp_path)
      datasets/gov/example              |              |         |            |
        | data                          | dask/memory  |         |            |
        |   |                           |              |         |            |
-       |   |   | Country               |              | _id     |            | open
+       |   |   | Country               |              |         |            | open
        |   |   |    | _id              | string       |         | _id        |
        |   |   |    | name             | string       |         | name       |
        |   |   |    |                  |              |         |            |